+++ packages/superset-ui-plugins-demo/storybook/stories/preset-chart-xy/ScatterPlot/index.ts
@@ -1,12 +1,14 @@
-import { ScatterPlotPlugin as LegacyScatterPlotPlugin } from '../../../../../superset-ui-preset-chart-xy/esm/legacy';
-import { ScatterPlotPlugin } from '../../../../../superset-ui-preset-chart-xy';
+import {
+  ScatterPlotChartPlugin,
+  LegacyScatterPlotChartPlugin,
+} from '../../../../../superset-ui-preset-chart-xy';
 import BasicStories from './stories/basic';
 import BubbleStories from './stories/bubble';
 import LegacyStories from './stories/legacy';
 import { SCATTER_PLOT_PLUGIN_TYPE, SCATTER_PLOT_PLUGIN_LEGACY_TYPE } from './constants';
 
-new LegacyScatterPlotPlugin().configure({ key: SCATTER_PLOT_PLUGIN_LEGACY_TYPE }).register();
-new ScatterPlotPlugin().configure({ key: SCATTER_PLOT_PLUGIN_TYPE }).register();
+new LegacyScatterPlotChartPlugin().configure({ key: SCATTER_PLOT_PLUGIN_LEGACY_TYPE }).register();
+new ScatterPlotChartPlugin().configure({ key: SCATTER_PLOT_PLUGIN_TYPE }).register();
 
 export default {
   examples: [...BasicStories, ...BubbleStories, ...LegacyStories],
