+++ packages/superset-ui-preset-chart-xy/src/components/Line/Line.tsx
@@ -14,11 +14,11 @@ import { WithLegend } from '@superset-ui/chart-composition';
 import { createSelector } from 'reselect';
 import { Dataset, PlainObject } from 'encodable/lib/types/Data';
 import DefaultTooltipRenderer from './DefaultTooltipRenderer';
-import createMarginSelector, { DEFAULT_MARGIN } from '../utils/createMarginSelector';
-import convertScaleToDataUIScale from '../utils/convertScaleToDataUIScaleShape';
-import createXYChartLayoutWithTheme from '../utils/createXYChartLayoutWithTheme';
-import createRenderLegend from '../components/legend/createRenderLegend';
-import { LegendHooks } from '../components/legend/types';
+import createMarginSelector, { DEFAULT_MARGIN } from '../../utils/createMarginSelector';
+import convertScaleToDataUIScale from '../../utils/convertScaleToDataUIScaleShape';
+import createXYChartLayoutWithTheme from '../../utils/createXYChartLayoutWithTheme';
+import createRenderLegend from '../legend/createRenderLegend';
+import { LegendHooks } from '../legend/types';
 import {
   lineEncoderFactory,
   LineEncoder,
