+++ packages/superset-ui-preset-chart-xy/src/components/ScatterPlot/DefaultTooltipRenderer.tsx
